# docs/openapi.yaml
openapi: '3.1.0'
info:
  title: Biblioteca API
  version: '1.0.0'
  description: |
    API REST do Sistema de Biblioteca.
    - Autenticação via JWT em cookie `httpOnly` (ADR-0003)
    - Dois papéis: `leitor` e `bibliotecario` (RN-7)
    - Terminologia segue `docs/produto/glossario.md`
  license:
    name: Privado
servers:
  - url: http://localhost:3000
    description: Desenvolvimento local
  - url: https://api.biblioteca.example.com
    description: Produção
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Access token OIDC emitido pelo **Keycloak** (RS256, 15 min) — ADR-0009.
        A API não emite token: obtenha-o no realm `biblioteca` e envie no header
        `Authorization`. Não há cookie de sessão.

        Discovery: `https://localhost:8443/realms/biblioteca/.well-known/openid-configuration`

        O token precisa ter `iss` do realm e `aud` contendo `biblioteca-api`.
    openId:
      type: openIdConnect
      openIdConnectUrl: https://localhost:8443/realms/biblioteca/.well-known/openid-configuration

  # -------------------------------------------------------------------------
  # Schemas — espelham packages/shared/src/types/domain.ts
  # -------------------------------------------------------------------------
  schemas:
...
    ReservationStatus:
      type: string
      enum: [active, expired, converted, cancelled]
      description: |
        Estado da Reserva, **derivado a cada leitura** — nunca persiste como coluna.
        `active` → dentro das 12h, sem nenhum desfecho (RN-1).
        `expired` → prazo vencido (RN-1), com ou sem o job já processado.
        `converted` → virou Empréstimo (RN-6). É um sucesso, não uma falha.
        `cancelled` → o **Leitor** desistiu antes do prazo (RF-L8, RN-11).

        `expired` e `cancelled` são desfechos distintos e não se confundem: um é
        esquecimento, o outro é desistência, e para o balcão a diferença explica
        por que a Cópia voltou antes da hora. Até a issue #20 os dois
        compartilhavam a coluna `cancelledAt` e este valor significava expiração.
...
    Book:
      type: object
      required: [id, isbn, title, genre, authorId, createdAt]
      properties:
        id:          { type: string }
        isbn:        { type: string, example: '9780156027748' }
        title:       { type: string, example: 'Ensaio sobre a Cegueira' }
        synopsis:
          oneOf:
            - type: string
            - type: 'null'
        genre:       { type: string, example: 'Ficção literária' }
        coverUrl:
          oneOf:
            - type: string
              format: uri
            - type: 'null'
        publishedAt:
          oneOf:
            - type: string
              format: date-time
            - type: 'null'
        authorId:    { type: string }
        createdAt:   { type: string, format: date-time }
...
    CopyRef:
      type: object
      description: |
        Projeção da Cópia embutida em Reserva e Empréstimo. **Não é uma `Copy`
        completa**: as respostas trazem apenas identificação e o Livro, sem
        `status` nem `bookId`. Um cliente que assumir `Copy` aqui lê `undefined`
        em campos que o schema prometeria como obrigatórios.
      required: [id, code, book]
      properties:
        id:   { type: string }
        code: { type: string, example: '9780156027748-C1' }
        book:
          allOf:
            - $ref: '#/components/schemas/Book'
            - type: object
              required: [author]
              properties:
                author:
                  type: object
                  required: [id, name]
                  properties:
                    id:   { type: string }
                    name: { type: string }
...
    Reservation:
      type: object
      required: [id, expiresAt, copyId, userId, createdAt]
      properties:
        id:          { type: string }
        expiresAt:   { type: string, format: date-time, description: 'createdAt + 12h (RN-1)' }
        convertedAt:
          oneOf:
            - type: string
              format: date-time
            - type: 'null'
        expiredAt:
          oneOf:
            - type: string
              format: date-time
            - type: 'null'
          description: |
            Quando o job de expiração (`expireReservationsTx`) registrou que o
            prazo de 12h venceu e devolveu a Cópia ao acervo (RN-1, RN-5). Fica
            `null` na janela de até 60 s entre o vencimento e a passagem do job —
            por isso `status` deriva de `expiresAt` também, e não só daqui.
        cancelledAt:
          oneOf:
            - type: string
              format: date-time
            - type: 'null'
          description: |
            Quando o **Leitor** cancelou a própria Reserva (RF-L8, RN-11).
            Exclusivo de desistência: expiração vive em `expiredAt`.

            Até a issue #20 este campo era também onde o job gravava a expiração,
            e a orientação era justamente **não** exibir "cancelada" a partir
            dele. Com os campos separados, exibir é o certo.
        copyId:      { type: string }
        userId:      { type: string }
        createdAt:   { type: string, format: date-time }
    ReservationDetail:
      allOf:
        - $ref: '#/components/schemas/Reservation'
        - type: object
          required: [copy, user]
          properties:
            copy:
              $ref: '#/components/schemas/CopyRef'
            status:
              allOf:
                - $ref: '#/components/schemas/ReservationStatus'
              description: |
                Enviado por `GET /reservations`. **Ausente em
                `GET /me/reservations`**, que já filtra no banco e devolve
                somente Reservas ativas — ali a ausência significa `active`.
            user:
              type: object
              required: [id, name, email]
              properties:
                id:    { type: string }
                name:  { type: string }
                email: { type: string, format: email }
...
    ApiError:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:    { type: string, example: 'RESERVATION_NOT_FOUND' }
            message: { type: string, example: 'Reserva não encontrada' }
    Pagination:
      type: object
      required: [page, pageSize, total, totalPages]
      properties:
        page:       { type: integer, minimum: 1 }
        pageSize:   { type: integer, minimum: 1, maximum: 100 }
        total:      { type: integer, minimum: 0 }
        totalPages: { type: integer, minimum: 0 }

  # -------------------------------------------------------------------------
  # Responses reutilizáveis
  # -------------------------------------------------------------------------
  responses:
    Unauthorized:
      description: |
        Não autenticado — token ausente, expirado, mal assinado, de outro
        emissor (`iss`) ou destinado a outra audiência (`aud`).
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
    Forbidden:
      description: Não autorizado — papel insuficiente (RN-7)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
          example:
            error:
              code: FORBIDDEN
              message: Apenas bibliotecários podem executar esta ação
...
paths:
...
  /reservations:
    post:
      summary: Criar reserva (RF-L3)
      description: |
        Leitor cria uma reserva para um livro com disponibilidade.
        Vincula automaticamente a uma Cópia disponível.
        A Reserva expira em 12h (RN-1). Requer papel `leitor`.

        Recusas por regra do Leitor, ambas 409: já ter Reserva ativa ou
        Empréstimo em aberto deste mesmo Livro (RN-9), e ter atingido o teto de
        Reservas ativas simultâneas (RN-10). Distinguem-se pelo `code`.
      operationId: createReservation
      tags: [Reservas]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [bookId]
              properties:
                bookId: { type: string }
      responses:
        '201':
          description: Reserva criada
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: object
                    required: [reservation]
                    properties:
                      reservation: { $ref: '#/components/schemas/ReservationDetail' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404':
          description: Livro não existe no acervo
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiError' }
              example:
                error:
                  code: NOT_FOUND
                  message: 'Livro não encontrado: ckv1a2b3c4d5e6f7g8h9'
        '409':
          description: |
            Três recusas distintas compartilham o status; o `code` separa:

            - `NO_COPY_AVAILABLE` — o Livro existe, mas nenhuma Cópia está
              disponível (RN-3). Também é a resposta de quem perde a disputa pela
              última Cópia entre dois pedidos simultâneos.
            - `DUPLICATE_RESERVATION` — o Leitor já tem Reserva ativa ou
              Empréstimo em aberto deste Livro (RN-9).
            - `RESERVATION_LIMIT_REACHED` — o Leitor atingiu o teto de Reservas
              ativas simultâneas (RN-10).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiError' }
              examples:
                semCopia:
                  summary: RN-3 — acervo sem Cópia livre
                  value:
                    error:
                      code: NO_COPY_AVAILABLE
                      message: Não há cópias disponíveis para este livro no momento.
                livroDuplicado:
                  summary: RN-9 — o Leitor já tem este Livro
                  value:
                    error:
                      code: DUPLICATE_RESERVATION
                      message: Você já tem uma reserva ativa ou um empréstimo em aberto deste livro.
                tetoAtingido:
                  summary: RN-10 — teto de Reservas ativas
                  value:
                    error:
                      code: RESERVATION_LIMIT_REACHED
                      message: >-
                        Você já tem 3 reservas ativas, o máximo permitido. Retire ou
                        aguarde a expiração de uma delas para reservar outro livro.
        '422':
          description: Corpo inválido (`bookId` ausente ou vazio)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiError' }
              example:
                error:
                  code: VALIDATION_ERROR
                  message: bookId obrigatório

    get:
      summary: Listar todas as reservas (RF-B1, RF-B3)
      description: |
        Lista reservas ativas de todos os leitores.
        Suporta filtro por `userId` (RF-B3). Requer papel `bibliotecario`.
      operationId: listReservations
      tags: [Reservas]
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: query
          description: Filtrar por leitor específico (RF-B3)
          schema: { type: string }
        - name: bookId
          in: query
          description: Filtrar por livro específico
          schema: { type: string }
        - name: page
          in: query
          schema: { type: integer, minimum: 1, default: 1 }
        - name: pageSize
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
      responses:
        '200':
          description: Lista de reservas ativas
          content:
            application/json:
              schema:
                type: object
                required: [data, pagination]
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/ReservationDetail' }
                  pagination:
                    $ref: '#/components/schemas/Pagination'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /reservations/{id}/cancel:
    patch:
      summary: Cancelar a própria reserva (RF-L8, RN-11)
      description: |
        O Leitor desiste de uma Reserva ativa **dele** e a Cópia volta ao acervo
        na mesma transação (RN-5) — sem esperar as 12h de RN-1. A vaga volta ao
        teto de RN-10 e o Livro volta a ser reservável por ele (RN-9 não conta
        Reserva cancelada). Requer papel `leitor`.

        Propriedade é regra, não filtro de conveniência: o Leitor sai do `sub` do
        token, nunca do corpo (ADR-0009), e a Reserva de outro responde **404**,
        não 403 — um 403 confirmaria a existência do id a quem não é dono.

        `PATCH`, e não `DELETE`, pelo mesmo motivo de `PATCH /loans/{id}/return`:
        a Reserva não é removida, ela recebe um desfecho e continua no registro.
      operationId: cancelReservation
      tags: [Reservas]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: |
            Reserva cancelada. Devolve a Reserva já encerrada (`status:
            cancelled`, `cancelledAt` preenchido) e não 204: a tela troca o
            rótulo da linha sem esperar o refetch da lista.
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: object
                    required: [reservation]
                    properties:
                      reservation: { $ref: '#/components/schemas/ReservationDetail' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403':
          description: |
            Papel `bibliotecario` (RN-2, RN-7). Cancelar é ato de quem reservou:
            o balcão desfaz Reserva convertendo ou esperando o prazo.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiError' }
        '404':
          description: |
            Reserva inexistente **ou de outro Leitor** — as duas respondem igual,
            de propósito (RN-11).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiError' }
              example:
                error:
                  code: NOT_FOUND
                  message: Reserva não encontrada.
        '409':
          description: |
            A Reserva já tem desfecho; o `code` separa o que fazer a seguir:

            - `CONFLICT` — já convertida em Empréstimo (o livro está com o
              Leitor: o caminho agora é Devolução no balcão), já cancelada
              (idempotência — dois cliques não são dois cancelamentos), ou
              corrida perdida para a efetivação no balcão.
            - `RESERVATION_EXPIRED` — o prazo de RN-1 chegou primeiro. É o caso
              comum: a aba ficou aberta e o botão continuou à mostra.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiError' }
              examples:
                jaConvertida:
                  summary: Já virou Empréstimo
                  value:
                    error:
                      code: CONFLICT
                      message: >-
                        Esta reserva já virou empréstimo e não pode ser cancelada.
                        Procure o balcão para devolver o livro.
                jaExpirada:
                  summary: RN-1 chegou primeiro
                  value:
                    error:
                      code: RESERVATION_EXPIRED
                      message: >-
                        Esta reserva expirou e a cópia já voltou ao acervo — não há
                        o que cancelar.
  /me/reservations:
    get:
      summary: Listar minhas reservas ativas (RF-L4)
      description: |
        Retorna apenas as Reservas ativas (não expiradas) do Leitor autenticado.
        Alvo: < 500 ms (RNF-4). Requer papel `leitor`.
      operationId: listMyReservations
      tags: [Reservas]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Reservas ativas do leitor
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/ReservationDetail' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }

  # -------------------------------------------------------------------------
  # Empréstimos — Bibliotecário (RF-B2, RF-B4, RF-B5) e Leitor (RF-L5)
  # -------------------------------------------------------------------------
...